refactor: move localeToDisplayInfo to its own function + remove useEffect for filtering
