+++ src/components/LanguagePicker/useLanguagePicker.tsx
@@ -1,33 +1,21 @@
-import { useEffect, useState } from "react"
+import { useMemo } from "react"
 import { useRouter } from "next/router"
 import { useTranslation } from "next-i18next"
 
-import type {
-  I18nLocale,
-  Lang,
-  LocaleDisplayInfo,
-  ProjectProgressData,
-} from "@/lib/types"
+import type { Lang, LocaleDisplayInfo } from "@/lib/types"
 
 import { MatomoEventOptions, trackCustomEvent } from "@/lib/utils/matomo"
-import { filterRealLocales, languages } from "@/lib/utils/translations"
+import { filterRealLocales } from "@/lib/utils/translations"
 
-import progressDataJson from "@/data/translationProgress.json"
-
-import { DEFAULT_LOCALE } from "@/lib/constants"
+import { localeToDisplayInfo } from "./localeToDisplayInfo"
 
 import useDisclosure from "@/hooks/useDisclosure"
-
-const progressData = progressDataJson satisfies ProjectProgressData[]
 
 export const useLanguagePicker = (handleClose?: () => void) => {
   const { t } = useTranslation("common")
   const { locale, locales: rawLocales } = useRouter()
 
-  const [filteredNames, setFilteredNames] = useState<LocaleDisplayInfo[]>([])
-
-  // perform all the filtering and mapping when the filter value change
-  useEffect(() => {
+  const languages = useMemo<LocaleDisplayInfo[]>(() => {
     const locales = filterRealLocales(rawLocales)
 
     // Get the preferred languages for the users browser
@@ -52,99 +40,26 @@ export const useLanguagePicker = (handleClose?: () => void) => {
     // Remove duplicate matches
     const browserLocales = Array.from(new Set(allBrowserLocales))
 
-    const localeToDisplayInfo = (localeOption: Lang): LocaleDisplayInfo => {
-      const i18nItem: I18nLocale = languages[localeOption]
-      const englishName = i18nItem.name
-
-      // Get "source" display name (Language choice displayed in language of current locale)
-      const intlSource = new Intl.DisplayNames([locale!], {
-        type: "language",
-      }).of(localeOption)
-      // For languages that do not have an Intl display name, use English name as fallback
-      const fallbackSource =
-        intlSource !== localeOption ? intlSource : englishName
-      const i18nKey = "language-" + localeOption.toLowerCase()
-      const i18nSource = t(i18nKey) // Falls back to English namespace if not found
-
-      // If i18nSource (fetched from `language-{locale}` in current namespace)
-      // is not translated (output === englishName), or not available
-      // (output === i18nKey), use the Intl.DisplayNames result as fallback
-      const sourceName = [i18nKey, englishName].includes(i18nSource)
-        ? fallbackSource
-        : i18nSource
-
-      // Get "target" display name (Language choice displayed in that language)
-      const fallbackTarget = new Intl.DisplayNames([localeOption], {
-        type: "language",
-      }).of(localeOption)
-      const i18nConfigTarget = i18nItem.localName
-      const targetName = i18nConfigTarget || fallbackTarget
-
-      if (!sourceName || !targetName) {
-        console.warn("Missing language display name:", {
-          localeOption,
-          sourceName,
-          targetName,
+    return (
+      (locales as Lang[])
+        ?.map((localeOption) => {
+          const displayInfo = localeToDisplayInfo(
+            localeOption,
+            locale as Lang,
+            t
+          )
+          const isBrowserDefault = browserLocales.includes(localeOption)
+          return { ...displayInfo, isBrowserDefault }
         })
-      }
-
-      // English will not have a dataItem
-      const dataItem = progressData.find(
-        ({ languageId }) =>
-          i18nItem.crowdinCode.toLowerCase() === languageId.toLowerCase()
-      )
-
-      const approvalProgress =
-        localeOption === DEFAULT_LOCALE
-          ? 100
-          : Math.floor(
-              (dataItem!.words.approved / dataItem!.words.total) * 100
-            ) || 0
-
-      const isBrowserDefault = browserLocales.includes(localeOption)
-
-      const returnData: Partial<LocaleDisplayInfo> = {
-        localeOption,
-        sourceName: sourceName ?? localeOption,
-        targetName: targetName ?? localeOption,
-        englishName,
-        isBrowserDefault,
-      }
-
-      if (progressData.length < 1) {
-        console.warn(`Missing translation progress data; check GitHub action`)
-        return {
-          ...returnData,
-          approvalProgress: 0,
-          wordsApproved: 0,
-        } as LocaleDisplayInfo
-      }
-
-      const totalWords = progressData[0].words.total
-
-      const wordsApproved =
-        localeOption === DEFAULT_LOCALE
-          ? totalWords || 0
-          : dataItem?.words.approved || 0
-
-      return {
-        ...returnData,
-        approvalProgress,
-        wordsApproved,
-      } as LocaleDisplayInfo
-    }
-
-    const displayNames: LocaleDisplayInfo[] =
-      (locales as Lang[])?.map(localeToDisplayInfo).sort((a, b) => {
-        const indexA = browserLocales.indexOf(a.localeOption as Lang)
-        const indexB = browserLocales.indexOf(b.localeOption as Lang)
-        if (indexA >= 0 && indexB >= 0) return indexA - indexB
-        if (indexA >= 0) return -1
-        if (indexB >= 0) return 1
-        return b.approvalProgress - a.approvalProgress
-      }) || []
-
-    setFilteredNames(displayNames)
+        .sort((a, b) => {
+          const indexA = browserLocales.indexOf(a.localeOption as Lang)
+          const indexB = browserLocales.indexOf(b.localeOption as Lang)
+          if (indexA >= 0 && indexB >= 0) return indexA - indexB
+          if (indexA >= 0) return -1
+          if (indexB >= 0) return 1
+          return b.approvalProgress - a.approvalProgress
+        }) || []
+    )
   }, [locale, rawLocales, t])
 
   const { isOpen, setValue, ...menu } = useDisclosure()
@@ -182,6 +97,6 @@ export const useLanguagePicker = (handleClose?: () => void) => {
   return {
     t,
     disclosure: { isOpen, setValue, onOpen, onClose },
-    filteredNames,
+    languages,
   }
 }
